Conflicts:
	cloudplatform-webportal-webadmin/src/main/webapp/WEB-INF/views/clouddb/hcluster_detail.jsp
	cloudplatform-webportal-webapp/src/main/webapp/WEB-INF/views/clouddb/db_detail.jsp
	pom.xml

+++ cloudplatform-webportal-data-model/src/main/java/com/letv/portal/python/service/impl/BuildTaskServiceImpl.java
@@ -640,7 +640,7 @@ public class BuildTaskServiceImpl implements IBuildTaskService{
 		Integer status = (Integer) ((Map)map.get("response")).get("status");
 		mcluster.setStatus(status);
 		this.mclusterService.updateBySelective(mcluster);
-		if(status == MclusterStatus.NOTEXIT.getValue()) {
+		if(status == MclusterStatus.NOTEXIT.getValue() && status == MclusterStatus.DESTROYED.getValue()) {
 			this.mclusterService.delete(mcluster);
 		}
 	}
@@ -652,7 +652,9 @@ public class BuildTaskServiceImpl implements IBuildTaskService{
 		Integer status = (Integer) ((Map)map.get("response")).get("status");
 		container.setStatus(status);
 		this.containerService.updateBySelective(container);
-		
+		if(status == MclusterStatus.NOTEXIT.getValue() && status == MclusterStatus.DESTROYED.getValue()) {
+			this.containerService.delete(container);
+		}
 	}
 	
 	public Integer transStatus(String statusStr){
